backend/app/core/security.py
# backend/app/core/security.py
import jwt
from datetime import datetime, timedelta
from typing import Optional
# The "." means "look in the current folder"
from .config import settings
# This matches the industry standard for signing tokens
ALGORITHM = "HS256"

def create_access_token(data: dict, expires_delta: Optional[timedelta] = None):
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES)
    
    to_encode.update({"exp": expire})
    # Secret key must be kept safe in your .env file
    encoded_jwt = jwt.encode(to_encode, settings.SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt