requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
asyncpg>=0.28.0
alembic>=1.11.1
redis>=4.6.0
langchain-core>=0.1.0
langgraph>=0.0.10
pydantic>=2.0.0
pydantic-settings>=2.0.0
qrcode[pil]>=7.4.2
httpx>=0.24.1
passlib[bcrypt]>=1.7.4
pyjwt>=2.8.0
python-multipart>=0.0.6
